refactor(windows)!: extract SDIO functionality into submodule
BREAKING CHANGE: Internal module structure changed
- SDIO code moved from windows.rs to windows/sdio.rs
- Public API remains unchanged via re-exports
- Improves code maintainability per roadmap goals
